requirements.txt
kafka-python==2.0.2
minio==7.2.5
elasticsearch==8.12.1
psycopg2-binary==2.9.9
redis==5.0.1
opencv-python==4.9.0.80
ultralytics==8.4.14
deep-sort-realtime==1.3.2
torch==2.1.0
torchvision==0.16.0
pillow==12.1.1
shapely==2.1.2
numpy==1.26.4
pandas==2.0.3
watchdog==4.0.0
docker==7.0.0
web3==6.15.1
requests==2.31.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dateutil==2.8.2
pytz==2023.3
python-dotenv==1.0.1
scapy==2.5.0
tqdm==4.66.2
colorama==0.4.6
pyyaml==6.0.1
scikit-learn==1.3.2
scipy==1.11.3
shap==0.43.0
joblib==1.3.2